requirements.txt
python-telegram-bot==20.3
PyPDF2==3.0.1
pikepdf==8.6.0
Pillow==10.0.1
reportlab==4.0.4
celery==5.3.0
redis==4.5.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
boto3==1.28.0   # for S3/MinIO
gunicorn==20.1.0
uvicorn==0.20.0
